Use $realpath_root instead of $document_root
Using `$document_root`, Nginx won't follow symlinks to actual file path. This will be a problem deploying application to the server in which we use symlinks to link application root (`root` in Nginx config) to latest release. It may worth mentioning that doing so will make make deployment smooth and add option to rollback to previous release by a single symlink update.