Remove example trailing slash middleware since it is not compatible
The example middleware to remove the trailing slash from routes in the
Slim V4 documentation is not PSR-15 compatible, and thus does not work
in Slim 4.  Thankfully, the linked middleweare is, so update the docs
to point solely to that middleware.

Additionally, clarify that rewriting a request of a non-GET operation
only affects the in-memory copy of the Request object, and not the
actual user called by the consumer.